Align chat stream fallback tests with SSE responses
